Working delete account and delete user record
Alerts before deletions
Some details with alerts UI

+++ digitagrizeit/code/deleteuser.php
@@ -4,10 +4,17 @@
 
     include 'connect.php';
     
-    $sql_delete = "DELETE FROM users WHERE userID='$userid'";
-    $delete = mysqli_query($conn, $sql_delete);
+    if(isset($_POST['deleteid'])){
 
-    $_GET['deleteok'] = "";
-    header("location: AdminPanel.php");
+        $userid = $_POST['deleteid'];
+        $sql_delete = "DELETE FROM users WHERE userID='$userid'";
+        $delete = mysqli_query($conn, $sql_delete);
+        echo 1;
+        
+        //header("location: AdminPanel.php");
+
+    }
+
+
 
 ?>
